requirements.txt
google-search-results>=2.4.2
openai>=1.0.0
rich>=13.7.0
python-dotenv>=1.0.0
markdown>=3.5
weasyprint>=60.0
